# skillhub-back/openapi.yaml
openapi: 3.0.3
info:
  title: API SkillHub
  version: 1.0.0
  description: Documentation de l'API du projet SkillHub (React + Laravel)
servers:
  - url: http://localhost:8000
paths:
...
  /api/login:
    post:
      summary: Connexion utilisateur
      description: Authentifie un utilisateur avec email et mot de passe. Retourne un token JWT.
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "formateur@test.com"
              password: "123456"
      responses:
        '200':
          description: Connexion réussie — token JWT retourné
          content:
            application/json:
              example:
                access_token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                token_type: "bearer"
                expires_in: 3600
                user:
                  id: 1
                  nom: "Dupont"
                  prenom: "Jean"
                  email: "formateur@test.com"
                  role: "FORMATEUR"
        '401':
          description: Email ou mot de passe incorrect
          content:
            application/json:
              example:
                error: "Email ou mot de passe incorrect"
        '500':
          description: Erreur serveur
  /api/register:
    post:
      summary: Inscription utilisateur
      description: Crée un nouveau compte utilisateur. Le rôle doit être APPRENANT, FORMATEUR ou ADMINISTRATEUR.
      requestBody:
        required: true
        content:
          application/json:
            example:
              nom: "Dupont"
              prenom: "Jean"
              email: "formateur@test.com"
              password: "123456"
              role: "FORMATEUR"
      responses:
        '201':
          description: Inscription réussie — compte créé et token retourné
          content:
            application/json:
              example:
                access_token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                token_type: "bearer"
                user:
                  id: 3
                  nom: "Dupont"
                  prenom: "Jean"
                  email: "formateur@test.com"
                  role: "FORMATEUR"
        '422':
          description: Erreur de validation — champs manquants ou email déjà utilisé
          content:
            application/json:
              example:
                message: "The email has already been taken."
                errors:
                  email:
                    - "The email has already been taken."
        '500':
          description: Erreur serveur
...